web_api/pyproject.toml
[project]
name = "hcc-plan-web-api"
version = "0.1.0"
description = "FastAPI web platform for hcc_plan"
requires-python = ">=3.12"
dependencies = [
    "hcc-plan",
    "apscheduler>=3.10",
    "bcrypt>=4.0",
    "fastapi>=0.115",
    "jinja2>=3.1",
    "pydantic-settings>=2.7",
    "PyJWT>=2.8",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.34",
]

[tool.uv.sources]
hcc-plan = {workspace = true}
